backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0
matplotlib>=3.9.0
plotly>=5.24.0
langchain>=0.3.10
langchain-mistralai>=1.1.0
langgraph>=0.2.60
pydantic>=2.10.0
aiofiles>=24.1.0
